# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Crispy N Snacky API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List active products
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: badge
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found

  # ── Customer Auth ─────────────────────────────────────────────────────────
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserRegisterInput" }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in as customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current customer
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Not authenticated
    put:
      operationId: updateMe
      tags: [auth]
      summary: Update customer profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdateInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }

  # ── Cart ──────────────────────────────────────────────────────────────────
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Admin staff login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                required: [staff]
                properties:
                  staff: { $ref: "#/components/schemas/StaffMember" }
        "401":
          description: Invalid credentials
...
  /admin/auth/me:
    get:
      operationId: adminGetMe
      tags: [admin-auth]
      summary: Get current staff member
      responses:
        "200":
          description: Current staff
          content:
            application/json:
              schema:
                type: object
                required: [staff]
                properties:
                  staff: { $ref: "#/components/schemas/StaffMember" }
        "401":
          description: Not authenticated

  # ── Admin Products ────────────────────────────────────────────────────────
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin-products]
      summary: List all products (including inactive)
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: adminCreateProduct
      tags: [admin-products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product: { $ref: "#/components/schemas/Product" }
  /admin/products/{id}:
    get:
      operationId: adminGetProduct
      tags: [admin-products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product: { $ref: "#/components/schemas/Product" }
    put:
      operationId: adminUpdateProduct
      tags: [admin-products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: adminDeleteProduct
      tags: [admin-products]
      summary: Soft-delete product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted

  # ── Admin Orders ──────────────────────────────────────────────────────────
...
  /admin/staff:
    get:
      operationId: adminListStaff
      tags: [admin-staff]
      summary: List all staff (owner only)
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: object
                required: [staff]
                properties:
                  staff:
                    type: array
                    items: { $ref: "#/components/schemas/StaffMember" }
    post:
      operationId: adminCreateStaff
      tags: [admin-staff]
      summary: Create staff member (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [staff]
                properties:
                  staff: { $ref: "#/components/schemas/StaffMember" }
  /admin/staff/{id}:
    put:
      operationId: adminUpdateStaff
      tags: [admin-staff]
      summary: Update staff member (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [staff]
                properties:
                  staff: { $ref: "#/components/schemas/StaffMember" }
    delete:
      operationId: adminDeleteStaff
      tags: [admin-staff]
      summary: Deactivate staff member (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deactivated
components:
  schemas:
...
    Product:
      type: object
      required: [id, slug, name, description, shortDescription, price, category, images, stock, gstRate, active, featured, createdAt, updatedAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        shortDescription: { type: string }
        price: { type: string }
        comparePrice: { type: ["string", "null"] }
        category: { type: string }
        images: { type: array, items: { type: string } }
        badge: { type: ["string", "null"] }
        stock: { type: integer }
        gstRate: { type: string }
        weight: { type: ["string", "null"] }
        active: { type: boolean }
        featured: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [slug, name, price, category]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        shortDescription: { type: string }
        price: { type: string }
        comparePrice: { type: ["string", "null"] }
        category: { type: string }
        images: { type: array, items: { type: string } }
        badge: { type: ["string", "null"] }
        stock: { type: integer }
        gstRate: { type: string }
        weight: { type: string }
        active: { type: boolean }
        featured: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        shortDescription: { type: string }
        price: { type: string }
        comparePrice: { type: ["string", "null"] }
        category: { type: string }
        images: { type: array, items: { type: string } }
        badge: { type: ["string", "null"] }
        stock: { type: integer }
        gstRate: { type: string }
        weight: { type: string }
        active: { type: boolean }
        featured: { type: boolean }
    User:
      type: object
      required: [id, email, name, active, createdAt, updatedAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        phone: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    UserRegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string }
        password: { type: string, minLength: 8 }
        name: { type: string }
        phone: { type: string }
    UserUpdateInput:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    StaffMember:
      type: object
      required: [id, email, name, role, active, createdAt, updatedAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string }
        active: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    StaffInput:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string }
        password: { type: string }
        name: { type: string }
        role: { type: string }
    StaffUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string }
        active: { type: boolean }
        password: { type: string }
...