# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Crispy N Snacky API
 * OpenAPI spec version: 0.2.0
 */

export interface ProductInput {
  slug: string;
  name: string;
  description?: string;
  shortDescription?: string;
  price: string;
  /** @nullable */
  comparePrice?: string | null;
  category: string;
  images?: string[];
  /** @nullable */
  badge?: string | null;
  stock?: number;
  gstRate?: string;
  weight?: string;
  active?: boolean;
  featured?: boolean;
}