# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Crispy N Snacky API
 * OpenAPI spec version: 0.2.0
 */

export interface StaffInput {
  email: string;
  password: string;
  name: string;
  role?: string;
}