# lib/api-zod/src/generated/types/userRegisterInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Crispy N Snacky API
 * OpenAPI spec version: 0.2.0
 */

export interface UserRegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
  phone?: string;
}